Add fallback console utilities when Rich is unavailable
